Added AGE to NetworkX function.
